// src/app/platform-service-api/model/synchronizedAccountAccessKey.ts
/**
 * EveKit Frontend Server
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 2.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { SynchronizedEveAccount } from './synchronizedEveAccount';


/**
 * EveKit synchronized account access key
 */
export interface SynchronizedAccountAccessKey {
    syncAccount?: SynchronizedEveAccount;
    /**
     * Unique key ID
     */
    kid?: number;
    /**
     * Key owner
     */
    account?: SynchronizedEveAccount;
    /**
     * Key name
     */
    keyName?: string;
    /**
     * Access key
     */
    accessKey?: number;
    /**
     * -1 if this key never expires, otherwise the date (milliseconds UTC) when this key expires
     */
    expiry?: number;
    /**
     * -1 if this key is unlimited, otherwise the date (milliseconds UTC) before which data may not be accessed
     */
    limit?: number;
    /**
     * Access credential
     */
    credential?: string;
    /**
     * Access key mask
     */
    maskValue?: number;
    /**
     * String value of access key mask
     */
    maskValueString?: string;
}