// src/app/sde-service-api/model/invType.ts
/**
 * EveKit SDE API Server
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 2.0.0.20180713
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface InvType {
    typeID?: number;
    groupID?: number;
    typeName?: string;
    description?: string;
    mass?: number;
    volume?: number;
    capacity?: number;
    portionSize?: number;
    raceID?: number;
    basePrice?: number;
    published?: boolean;
    marketGroupID?: number;
    iconID?: number;
    soundID?: number;
    graphicID?: number;
}